// apps/amplify-be/src/graphql/API.ts
export enum TransferStatus {
  COMPLETED = "COMPLETED",
  FAILED = "FAILED",
  REJECTED = "REJECTED",
  PENDING = "PENDING",
}
// ...
export type Transfer = {
  __typename: "Transfer",
  id: string,
  from: string,
  to: string,
  amount: number,
  status: TransferStatus,
  createdAt: string,
  updatedAt: string,
};